+++ lib/src/mask/mobile.dart
@@ -1,0 +1,7 @@
+/// 脱敏手机号
+String maskMobile(String mobile) {
+  if (mobile.length == 11) {
+    return "${mobile.substring(0, 3)}****${mobile.substring(7)}";
+  }
+  return mobile;
+}